requirements.txt
Django==4.2.14
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
